22.4.Webhooks: Preventing Replay Attacks{Update the Test}
To fix the test, we need to set a little bit of randomness to the event ID by adding a %s at the end and adding an mt_rand() to the sprintf()
Now, every event ID will be unique. Try the test again:

./bin/phpunit --filter testStripeCustomerSubscriptionDeleted
Green and happy!

Ok, enough webhooks. Let's do something fun, like making it possible for a user to upgrade from one subscription to another.