23.1.Upgrading Subscription Plans: The UI{Printing the Current Plan}
Imagine this: a sheep customer is so happy with the Farmer Brent subscription that they want to upgrade to the New Zealander! Awesome! Amazing! But also... currently impossible.
Time to fix that. Plan upgrades can be complex, because someone needs to calculate how much of the current month has been used and prorate funds towards the upgrade. Stripe's documentation talks about this. I'll guide you through everything, but this section is worth a read.

Printing the Current Plan
Buy a new Farmer Brent Subscription, and then head to the account page.
Let's focus on the plan upgrade user interface first. Here, I need to see which plan I'm currently subscribed to and a button to upgrade to the other plan.
Open the account.html.twig template and ProfileController.
Add a new variable: $currentPlan = null. Then, only if $this->getUser()->hasActiveSubscription(), set $currentPlan = $this->subscriptionHelper->findPlan() passing that $this->getUser()->getSubscription()->getStripePlanId(). The findPlan() method will give us a fancy SubscriptionPlan object. Pass a new current_plan variable into the template, set to $currentPlan. Then in the template, find the "Active Subscription" spot, and print current_plan.name.
Refresh the page! Great! Step 1 done: we have the "Farmer Brent" plan.