25.3.Upgrade: Processing the Upcoming Invoice{Charging Immediately for an Upgrade}
Finishing up the JS
Ok! It's time to execute this upgrade! To save us some time, I'll paste some JavaScript into the AJAX success function.
This first display how much we will charge the user. And check this out: it could be positive, meaning we'll charge them, or negative for a downgrade, meaning they'll get an account credit that will automatically be used for future charges.
Finally, this shows the user one last alert to confirm the change. If they click "Ok", the last callback will be executed. And it'll be our job to send one more AJAX call back to the server to finally change their plan.