26.1.Execute the Plan Upgrade{Changing a Subscription Plan in Stripe}
When the user clicks "OK", we'll make an AJAX request to the server and then tell Stripe to actually make the change.
In ProfileController, add the new endpoint: public function changePlanAction(). Set its URL to /profile/plan/change/execute/{planId} and name it account_execute_plan_change. Add the $planId argument. This will start just like the previewPlanChangeAction() endpoint: copy its $plan code and paste it here.

Changing a Subscription Plan in Stripe
To actually change the plan in Stripe, we need to fetch the Subscription, set its plan to the new id, and save. Super easy!
Open StripeClient and add a new function called changePlan() with two arguments: the User who wants to upgrade and the SubscriptionPlan that they want to change to.
Then, fetch the \Stripe\Subscription for the User with $this->findSubscription() passing it $user->getSubscription()->getStripeSubscriptionId(). Now, update that: $stripeSubscription->plan = $newPlan->getPlanId(). Finally, send that to Stripe with $stripeSubscription->save()
Ok, that was easy. And now you probably expect there to be a "catch" or a gotcha that makes this harder. Well... yea... there totally is. Sorry.
I told you earlier that Stripe doesn't charge the customer right now: it waits until the end of the cycle and then bills for next month's renewal, plus what they owe for upgrading this month. We want to bill them immediately.