26.2.Execute the Plan Upgrade{But Charge the User Immediately}
But Charge the User Immediately
I told you earlier that Stripe doesn't charge the customer right now: it waits until the end of the cycle and then bills for next month's renewal, plus what they owe for upgrading this month. We want to bill them immediately.
How? Simple: by manually creating an Invoice and paying it. Remember: when you create an Invoice, Stripe looks for all unpaid invoice items on the customer. When you change the plan, this creates two new invoice items for the negative and positive plan proration. So if we invoice the user right now, it will pay those invoice items.
And hey! We already have a method to do that called createInvoice(). Heck it even pays that invoice immediately. In our function, call $this->createInvoice() and pass it $user. Finally, return $stripeSubscription at the bottom - we'll need that in a minute. Back in the controller, call this with $stripeSubscription = $this->stripeClient then ->changePlan($this->getUser(), $plan)