27.2.Failing Awesomely When Payments Fail{Failing Gracefully; Reverting the Plan without Proration}
Failing Gracefully
This whole mess starts in StripeClient, when we call $this->createInvoice(), because this might fail. Scroll up to that method.
In addition to calling this to upgrade a subscription plan, we also call this at checkout, even if there is no subscription. The problem is that if payment fails on checkout, this invoice will still exist, and Stripe will try again to charge it. Imagine having your card be declined at checkout, only for the vendor to try to charge it again later, without you ever having finished the checkout process!
Here's our rescue plan: if paying the invoice fails, we need to close it. By doing that, Stripe will not try to pay it again.
To do that, surround the pay() line with a try-catch for the \Stripe\Exception\CardException exception. Here, add $invoice->close = true and then $invoice->save(). Then, re-throw the exception.
Our checkout logic looks for this exception and uses it to notify the user of the problem.
Next, down in the other function, if we fail to create the invoice, we need to not change the customer's plan in Stripe. Add a new variable called $originalPlanId set to $stripeSubscription->plan->id.
Then, surround the createInvoice() call with a try-catch block for the same exception: \Stripe\Exception\CardException

Reverting the Plan without Proration
If this happens, we need to do change the subscription plan back to the original one: $stripeSubscription->plan = $originalPlanId. But here's the tricky part: add $stripeSubscription->prorate = false.
Why? When we originally change the plan, that creates the two proration invoice items. If the invoice fails to pay, then the invoice containing those invoice items is closed. And that means, effectively, those invoice items have been deleted, which is good! In fact, it's exactly what we want.
But when we change from the new plan back to the old plan, we don't want two new proration invoice items in reverse to be created. By saying prorate = false, we're telling Stripe to change back to the original plan, but without creating any invoice items. Yep, simply change the plan back.
Finally, call $stripeSubscription->save(). Then, once again, re-throw the exception.
That fixes the problem in Stripe.