29.2.Monthly to Yearly: The Billing Period Change{Changing Duration changes Billing Period}
Changing Duration changes Billing Period
So if you're normally billed on the first of the month and you change from monthly to yearly on the 15th, you'll be credited half of your monthly subscription and then charged for a full year. That yearly subscription will start immediately, on the 15th of that month and be renewed in one year, on the 15th.
For us, this means that the amount_due on the Invoice is actually correct: we don't need to adjust it. In ProfileController, create a new variable called $currentUserPlan set to $this->subscriptionHelper->findPlan() and pass it $this->getUser()->getSubscription()->getStripePlanId().
Now, if $plan - which is the new plan - $plan->getDuration() matches the $currentUserPlan->getDuration(), then we should correct the total. Otherwise, if the duration is changing, the $total is already perfect.
Since this looks totally weird, I'll tweak my comment to mention that amount_due contains the extra month charge only if the duration stays the same.
Ok! Go back and refresh! Click "Bill yearly". Yes! That looks right: $891.06.
Because of this behavior difference when the duration changes, we need to fix one other spot: in StripeClient::changePlan(). Right now, we manually create an invoice so that the customer is charged immediately. But... we don't need to do that in this case: Stripe automatically creates and pays an Invoice when the duration changes.
In fact, trying to create an invoice will throw an error! Let's see it. First, update your credit card to one that will work.
Now, change to Bill yearly and confirm. The AJAX call should fail... and it does! Open the profiler for that request and find the Exception:

  Nothing to invoice for customer