29.3.Monthly to Yearly: The Billing Period Change{Duration Change? Don't Invoice}
Duration Change? Don't Invoice

Obviously, we need to avoid this. In StripeClient, add a new variable: $currentPeriodStart that's set to $stripeSubscription->current_period_start.
That's the current period start date before we change the plan.
After we change the plan, if the duration is different, the current period start will have changed. Surround the entire invoicing block with if $stripeSubscription->current_period_start == $currentPeriodStart. In other words: only invoice the customer manually if the subscription period hasn't changed. I think we should add a note above this: this can look really confusing!

Take it for a Test Drive
But, now it should work! Reset things by going to the pricing page and buying a brand new monthly subscription. Now, head to your account page and update it to yearly. The amount - $891 - looks right, so hit OK.
Yes! Plan changed! My option changed to "Bill monthly" and the "Next Billing at" date is August 10th - one year from today. We should probably print the year.
In Stripe, under payments, we have one for $891, and the customer is on the Farmer Brent yearly plan.