31.1.Validate that Coupon in Stripe! {Fetching the Coupon Information}
To fetch that coupon POST parameter, add the Request object as an argument. Then add $code = $request->request->get('code'). And in case some curious user submits while the form is empty, send back a validation message with $this->addFlash('error', 'Missing coupon code'). Redirect to the checkout page.

Fetching the Coupon Information
Great! At this point, all we need to do is talk to Stripe and ask them:
  Hey Stripe! Is this a valid coupon code in your system?. Oh, and if it is, ow much is it for?

Since Coupon is just an object in Stripe's API, we can fetch it like anything else. Booya!
As usual, add the API call in StripeClient. At the bottom, create a new public function called findCoupon() with a $code argument. Then, return \Stripe\Coupon::retrieve() and pass it the $code string, which is the Coupon's primary key in Stripe's API.
Back in OrderController, add $stripeCoupon = $this->stripeClient and then call ->findCoupon($code). If the code is invalid, Stripe will throw an exception. We'll handle that in a few minutes. But just for now, let's dump($stripeCoupon) and die to see what it looks like.
Ok, refresh, hit "I have a coupon code," fill in our CHEAP_SHEEP code, and submit!
There it is! In the _values section where the data hides, the coupon has an id, it shows the amount_off in cents and has a few other things, like duration, in case you want to create coupons that are recurring and need to tell the user that this will be applied multiple times.