31.2.Validate that Coupon in Stripe! {Fetching the Coupon Information}
Now that we know the coupon is legit, we should add it to our cart. I've already prepped the cart to be able to store coupons. Just use $this->shoppingCart and then call ->setCouponCode(), passing it the $code string and the amount off, in dollars: so $stripeCoupon->amount_off/100.
The cart will remember - via the session - that the user has this coupon.
We're just about done: add a sweet flash message - "Coupon applied!" - and then redirect back to the checkout page.
Refresh and re-POST the form! Coupon applied! Except... I don't see any difference: the total is still $99.