31.3.Validate that Coupon in Stripe! {Showing the Code on Checkout}
Showing the Code on Checkout
Refresh and re-POST the form! Coupon applied! Except... I don't see any difference: the total is still $99.
Here's why: it's specific to our ShoppingCart object. In checkout.html.twig, we print cart.total.
I designed the ShoppingCart class so that the getTotal() method adds up all of the product prices plus the subscription total.
But, this method doesn't subtract the coupon discount. I did this to keep things clean: total is really more like a "sub-total".
But no worries, the method below this - getTotalWithDiscount() - subtracts the coupon code:

 155 lines  src/Store/ShoppingCart.php
class ShoppingCart
{
    public function getTotalWithDiscount()
    {
        return max($this->getTotal() - $this->getCouponCodeValue(), 0);
    }
}
So back in the template, use cart.totalWithDiscount.
Ah, now it shows $49.

But, it'll be even clearer if we display the discount in the table. At the bottom of that table, add a new if statement: if cart.couponCode and an endif. Then, copy the subscription block from above, paste it here, and change the first variable to cart.couponCode and the second to cart.couponCodeValue without the / month, unless you want to make all your coupons recurring. Oh, and add "Coupon" in front of the code.
This time, the whole page makes sense! $99 - $50 = 9. It's a miracle!
Now for the easy step: apply the coupon to the user's order at checkout... ya know, so that they actually save $50.