32.3.Applying a Coupon at Checkout{Expired Coupons}
Expired Coupons
There's just one other situation to handle. In Stripe, find the Coupon section and create a second code. Set the amount to 0, duration once and the code: SINGLE_USE. By here's the kicker: set Max redemptions to 1. So, only one customer should be able to use this. There's also a time-sensitive "Redeem by" option.
Quickly, go use the SINGLE_USE code and fill out the form to checkout. This will be the first - and only - allowed "redemption" of this code. When you refresh the Coupon page in Stripe, Redemptions are 1/1.
Now, add another subscription to your cart. If you tried to use the code a second time, our system would allow this. And that makes sense: all we're doing now is looking up the code in Stripe to make sure it exists.
But, if we tried to checkout, Stripe would be pissed: it would not allow us to use the code a second time. Stripe has our back.
But, we should definitely prevent the code from being attached to the cart in the first place. Checkout the Coupon section of Stripe's API docs. Ah, this valid field is the key. This field basically answers this question:

  In this moment, can this coupon be used?