33.3.Free (Ice Cream) Checkout! {Handling Free Checkout; Optionally Apply the Stripe Token}
Handling Free Checkout
But that's not quite everything we need to do. Thanks to the discount, Stripe will already know that it doesn't need to charge the user, and so, the customer doesn't need to have a card. But, the funny thing is, up until now, our OrderController logic is expecting a stripeToken to always be submitted:

src/Controller/OrderController.php
class OrderController extends BaseController{
    public function checkoutAction(Request $request){
        if ($request->isMethod('POST')) {
            $token = $request->request->get('stripeToken');
        }
    }
}
Remember that's the token that we get back from Stripe, after submitting the credit card information to them. We then pass that to chargeCustomer() and attach it to the Customer:

src/Controller/OrderController.php
class OrderController extends BaseController{
    private function chargeCustomer($token){
        if (! $user->getStripeCustomerId()) {
            $stripeCustomer = ->createCustomer($user, $token);
        } else {
            $stripeCustomer = $stripeClient->updateCustomerCard($user, $token);
        }
    }
}

Optionally Apply the Stripe Token
But now, our code needs to be smart enough to not try to attach the token to the Customer for free orders. At the top, add a sanity check: if there is no token, and the shopping cart's total with discount is not free... well, we have a problem! Throw a clear exception: the order is non-free... but we're missing the payment token! Next, when we call createCustomer(), we pass in the $token. Open StripeClient and find that method. Hmm. Now, $paymentToken might be blank. But Stripe will be really angry if we try to attach an empty source to the Customer. Instead of doing this all at once, add a new $data array variable, and move the email key into it. Then, pass $data to the create() call. You know what's next: if $paymentToken is not blank, add a source key to $data set to $paymentToken.
We're done here.