33.4.Free (Ice Cream) Checkout! {Optionally Apply the Stripe Token}
But we have the same problem in updateCustomerCard(). Let's fix this here: if $stripeToken, then update the customer's card. Else, we do need to fetch the \Stripe\Customer object - we use it below. In StripeClient, add a new method to do this: public function findCustomer() with a User argument.
Then, return the timeless \Stripe\Customer::retrieve($user->getStripeCustomerId()). In the controller, call that: $stripeCustomer = $stripeClient->findCustomer($user). Ok, I'm feeling good! The last trouble spot is in updateCardDetails(). Open SubscriptionHelper.
Oh yea - this method looks at the sources key on the Customer to get the card brand and last four digits. In our app, every Customer has exactly one card, so we use the 0 key. But guess what! Not anymore: a customer might have zero cards.
So we just need to code defensively: add an if statement: if ! $stripeCustomer->sources->data, just return: there's no card details to update.
Ok, we're done! A free checkout and a normal checkout are almost the same. The only difference is that you don't have a Stripe token, so you can't attach that to your Customer.
Refresh the checkout page and, "Checkout for Free".
It works! In Stripe, find our Customer. There is no new payment, but there is an Invoice for $0 and an active subscription. The Invoice shows off the discount.