33.5.Free (Ice Cream) Checkout! {No Card? Webhook Problems}
No Card? Webhook Problems
Thanks to this change, it's now possible for a Customer to not have any cards attached in Stripe. And yea know what? This creates a new problem in a totally unrelated part of the process: webhooks.
But, it's no big deal. Open WebhookController and find the invoice.payment_failed section. Wait! Woh! Before that - oh geez - fix my horrible typo: invoice.payment_succeeded.
This is why you must test your webhooks!
Anyways, back to invoice.payment_failed. Our entire reason for handling this webhook is so that we can send the user an email to tell them that we're having a problem charging their card. We didn't actually do the work, but that email would probably sound like this:
  Hey friend! So, we're having a problem charging your card. If you need to update it, go to your account page and add the new details there.

But what will happen if a user checks out for free, but with a subscription? After their first month, Stripe will not be able to charge them, and it will trigger this webhook.
In those cases, the email should really have some different text, like:

  Yo amigo! I hope you enjoyed your free month. If you want to continue, you can add a credit card to your account page.
So to know which language to use, first fetch the Stripe Customer by saying $this->stripeClient->findCustomer($user). Now we can create a new variable, called $hasCardOnFile. Set that to a count of $stripeCustomer->sources->data and check if it's greater than zero. Now, you can use that variable to write the most uplifting, majestic, and encouraging payment failed emails that the world has ever seen.