34.2.Sweet Invoices{Listing all Invoices}
Listing all Invoices
Thanks to findPaidInvoices function, on the account page, at the bottom, we'll print a list of all the Customer's invoices. Eventually, they'll be able to click each invoice to see all the details.
Start in ProfileController... all the way at the top: this method renders the account page. Fetch the invoices with $invoices = $this->stripeClient->findPaidInvoices() with the current user. Pass that as a new variable into the template. Now inside the template, find the bottom of the table. Add a new row, and title it Invoices.
Next, create a list and then loop with for invoice in invoices. Add the endfor. Create an anchor tag, but keep the href empty for now - we don't have an invoice "show" page yet. Add some classes to make this look a little fancy.
So let's see, the user might be looking for a specific invoice, so let's print its date. Check out the Invoice API. Hey! There's actually a date field, which is a UNIX timestamp. Print invoice.date and then pipe that through the date filter with Y-m-d.
Next, add a span label, float it right, and inside, add the amount: $ then print invoice.amount_due / 100 to convert it from cents to dollars.
The amount_due field is what the user should have actually been charged, after accounting for coupons or a positive account balance.
Try things out so far: head back to the account page and refresh! Bam! Here's our long invoice list. Next, let's give each invoice its own detailed display page, complete with invoice items, discounts and anything else that might have happened on that invoice.